pyproject.toml
[tool.poetry]
name = "world-glimpse-front"
version = "0.1.0"
description = ""
authors = ["Marta Rg <martaroguerra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.0"
streamlit = "^1.26.0"
geopandas = "^0.13.2"
keplergl = "^0.3.2"
matplotlib = "^3.7.2"
streamlit-keplergl = "^0.3.0"
altair = "^5.1.1"
plost = "^0.2.5"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
